make !isequal('x', 120) and eprecate Int-Char comparisons, 'x' == 120
See https://github.com/JuliaLang/julia/issues/15983#issuecomment-213038173

+++ base/datafmt.jl
@@ -23,7 +23,7 @@ function countlines(io::IO, eol::Char='\n')
     while !eof(io)
         nb = readbytes!(io, a)
         @simd for i=1:nb
-            @inbounds nl += a[i] == eol
+            @inbounds nl += a[i] == UInt8(eol)
         end
     end
     nl
